Final version of Tier 3 Red Wine Regression Case Study
Case Study findings used Ordinary Least Squares Regression algorithm. The best model was the fourth model, which used "residual.sugar", "chlorides", "total.sulfur.dioxide", "density", and "sulphates" as predictors for pH. It was the best model in large part because we removed "citric acid" and "acid volatility" to avoid redundancies when predicting "pH".